Remove support for clang < 16
We are starting to support Flang on a separated branch and
that only works on clang 17+, but clang 17 is still "too new"
and is not really available on every distro.

But we can help ourselves by removing support for clang 13, 14 and 15
already.

With this I also disabled the FreeBSD CI because they don't have llvm16
yet