Speed up write operations
Unfortunately I had to change the tables of the database to
speedup some operations here - I appreciate a good review.

This uses the `on conflict do nothing` SQlite instructions
to not fail if we are adding something that already exists
on a database, instead of querying if things exists and then
adding if they don't.

This reduced the total time of saving the database from
~2 minutes on a distributed db merge to 1.50 - 10 seconds reduction
for merging 26 database files.

Signed-off-by: Tomaz Canabrava <tcanabrava@kde.org>